+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Fixed
-- Description message does not leave artifacts if there are no details
+- Description messages for pass and fail do not leave artifacts if there are no details
 
 ## [0.4.0] - 2016-11-10
 ### Added